+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpIsoProtocolTest.kt
@@ -77,9 +77,7 @@ class OpenId4VpIsoProtocolTest : FreeSpec({
             walletUrl,
             RequestOptions(
                 credentials = setOf(
-                    RequestOptionsCredential(
-                        MobileDrivingLicenceScheme, ISO_MDOC, setOf(GIVEN_NAME)
-                    )
+                    RequestOptionsCredential(MobileDrivingLicenceScheme, ISO_MDOC, setOf(GIVEN_NAME))
                 )
             ),
             holderOid4vp